refact: move ProfileSearcher to core.profile.search
